Implemented resource_permission API, added initial permission push to DB, more changes
