+++ son-gtkusr/helpers/init.rb
@@ -27,4 +27,4 @@
 
 require_relative 'helpers'
 require_relative 'keycloak'
-# require_relative 'authorization'
+require_relative 'authorization'
